pyproject.toml
[project]
name = "servicekit"
version = "0.1.0"
description = "Async SQLAlchemy framework with FastAPI integration - reusable foundation for building data services"
readme = "README.md"
authors = [{ name = "Morten Hansen", email = "morten@winterop.com" }]
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.13,<3.14"
keywords = [
  "fastapi",
  "sqlalchemy",
  "async",
  "database",
  "rest-api",
  "crud",
  "framework",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Framework :: FastAPI",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Database",
]
dependencies = [
  "aiosqlite>=0.21.0",
  "alembic>=1.17.0",
  "fastapi[standard]>=0.119.0",
  "gunicorn>=23.0.0",
  "opentelemetry-api>=1.37.0",
  "opentelemetry-exporter-prometheus>=0.58b0",
  "opentelemetry-instrumentation-fastapi>=0.58b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.58b0",
  "opentelemetry-sdk>=1.37.0",
  "pandas>=2.3.3",
  "pandas-stubs>=2.3.2.250926",
  "pydantic>=2.12.0",
  "python-ulid>=3.1.0",
  "sqlalchemy[asyncio]>=2.0.43",
  "structlog>=24.4.0",
]

[project.urls]
Homepage = "https://github.com/winterop-com/servicekit"
Repository = "https://github.com/winterop-com/servicekit"
Issues = "https://github.com/winterop-com/servicekit/issues"
Documentation = "https://winterop-com.github.io/servicekit"

[dependency-groups]
dev = [
  "coverage[toml]>=7.6.0",
  "mypy>=1.18.2",
  "pytest>=8.4.2",
  "pytest-cov>=5.0.0",
  "pytest-asyncio>=1.2.0",
  "ruff>=0.14.0",
  "pyright>=1.1.406",
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.0",
  "mkdocstrings[python]>=0.26.0",
]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "W", "F", "I", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"**/__init__.py" = ["D104"]
"src/**/*.py" = ["D102", "D105", "D107"]
"examples/**/*.py" = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["examples", ".git", ".venv", "__pycache__"]

[tool.coverage.run]
branch = true
dynamic_context = "test_function"
relative_files = true
source = ["servicekit"]

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:"]
precision = 2
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_unused_ignores = true
strict_equality = true
mypy_path = ["src", "typings"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pyright]
include = ["src", "tests", "examples"]
pythonVersion = "3.13"
typeCheckingMode = "strict"
diagnosticMode = "workspace"
useLibraryCodeForTypes = true
reportPrivateUsage = false
reportUnusedFunction = false
reportUnusedTypeParameters = true
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
